Allow adding sales tax to the amount with a checkbox
I am currently in the US and whenever I buy something the price is net.
Without this change I have to multiply the amount by 1.0875 manually for
every partial transaction as the receipts usually only show the sales
tax for the total.
Now I can just tick a checkbox to add the sales tax.
The sales tax percentage is saved and can be edited by pressing the
button next to the checkbox.